remove ContainableLayout
